LayoutTests/editing/selection/ios/do-not-show-selection-in-empty-borderless-subframe-expected.txt

This test verifies that selection UI is suppressed when focusing an element inside an empty subframe. To run the test manually, tap the button and check that a blue selection caret is not visible.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS caretRect.width is 0
PASS caretRect.height is 0
PASS successfullyParsed is true

TEST COMPLETE
Focus the hidden frame
